refactor: 컴포넌트 디렉토리 이동 및 텍스트 카드 디자인 수정
