+++ client/src/components/Ranking/MainRanking.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import Link from "next/link";
-import styles from "./MainRanking.module.css";
+import styles from "./MainRanking.module.scss";
 import {
   Box,
   Typography,
