+++ client/src/components/Result/TextCard.module.scss
@@ -1,0 +1,20 @@
+.box {
+  width: 200px;
+  height: 40px;
+  background-color: #FFC300;
+  color: #604900;
+  border: 1px solid #604900;
+  border-radius: 4px;
+  margin: 30px 0 10px 0;
+  text-align: center;
+  &:hover {
+    cursor: pointer;
+    background-color: #604900;
+    color: white;
+  }
+}
+
+.text {
+  line-height: 40px;
+  margin: auto 0;
+}
