+++ client/src/components/Search/MainSearch.module.scss
