+++ client/src/components/Search/SearchBar.module.scss
