+++ client/src/components/Search/SearchBar.tsx
