+++ client/src/components/Search/SearchImageBox.module.scss
