+++ client/src/components/Search/SearchImageBox.tsx
